MyFridgeVO, ingredientDTO, ingredient 클래스 생성 및 수정
